requirements.txt
streamlit 
pandas 
python-dotenv 
langchain 
langchain-google-genai 
pydantic 
google-generativeai